requirements.txt
fastapi==0.63.0
uvicorn==0.13.2
httpx==0.17.1
pydantic[email]==1.7.3
motor==2.3.1
python-decouple==3.4
python-jose[cryptography]==3.2.0